feat: Web API AI 端点真实数据集成 + clippy warning 清零
- AI 路由集成: 12 个 handler 从硬编码桩数据改为真实数据源
  (DecisionRepository/AuditLogger/OtaManager/OnlineUpdater/AbTestManager)
- 新增 AbTestManager 内存 CRUD、RequireRole 认证提取器
- Router 类型修复: Router<Arc<AppState>> + .with_state()
- DecisionRepository 补充 get_by_id 方法
- Clippy warning 清零: 59 个 warning 修复 (17 crates)
  (Default derive、needless borrow、scoped MutexGuard、vec![]
   替换 vec_init_then_push、range→iterator 等)

+++ mupc/crates/iec61850-plugin/src/goose.rs
@@ -183,7 +183,7 @@ impl GooseSubscriber {
     /// 检查 GOOSE 报文是否有效
     pub fn validate_goose(&self, msg: &GooseMessage) -> bool {
         // 检查 AppID 和 GOID 匹配
-        msg.app_id == self.config.app_id as u32 && msg.go_id == self.config.go_id
+        msg.app_id == self.config.app_id && msg.go_id == self.config.go_id
     }
 }
